+++ src/client/java/com/theohenno/narc_mod/entities/client/DroneEntityRenderState.java
@@ -1,0 +1,11 @@
+package com.theohenno.narc_mod.entities.client;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.render.entity.state.LivingEntityRenderState;
+
+@Environment(EnvType.CLIENT)
+public class DroneEntityRenderState extends LivingEntityRenderState {
+    public DroneEntityRenderState() {
+    }
+}
