+++ src/client/java/com/theohenno/narc_mod/entities/client/DroneEntityRenderer.java
@@ -1,0 +1,26 @@
+package com.theohenno.narc_mod.entities.client;
+
+import com.theohenno.narc_mod.NarcMod;
+import com.theohenno.narc_mod.entities.DroneEntity;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.render.entity.EntityRendererFactory;
+import net.minecraft.client.render.entity.MobEntityRenderer;
+import net.minecraft.util.Identifier;
+
+@Environment(EnvType.CLIENT)
+public class DroneEntityRenderer extends MobEntityRenderer<DroneEntity, DroneEntityRenderState, DroneEntityModel> {
+    public DroneEntityRenderer(EntityRendererFactory.Context context) {
+        super(context, new DroneEntityModel(context.getPart(DroneModelLayer.DRONE)), 0.5f);
+    }
+
+    @Override
+    public Identifier getTexture(DroneEntityRenderState state) {
+        return Identifier.of(NarcMod.MOD_ID, "textures/entity/drone/drone.png");
+    }
+
+    @Override
+    public DroneEntityRenderState createRenderState() {
+        return new DroneEntityRenderState();
+    }
+}
