+++ src/client/java/com/theohenno/narc_mod/screens/DroneScreen.java
@@ -1,0 +1,28 @@
+package com.theohenno.narc_mod.screens;
+
+import com.theohenno.narc_mod.entities.screen_handler.DroneScreenHandler;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.gui.screen.ingame.HandledScreen;
+import net.minecraft.client.render.RenderLayer;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.text.Text;
+import net.minecraft.util.Identifier;
+
+@Environment(EnvType.CLIENT)
+public class DroneScreen extends HandledScreen<DroneScreenHandler> {
+    private static final Identifier TEXTURE = Identifier.ofVanilla("textures/gui/container/chest.png");
+
+    public DroneScreen(DroneScreenHandler handler, PlayerInventory inventory, Text title) {
+        super(handler, inventory, title);
+        ++this.backgroundHeight;
+    }
+
+    @Override
+    protected void drawBackground(DrawContext context, float deltaTicks, int mouseX, int mouseY) {
+        int i = (this.width - this.backgroundWidth) / 2;
+        int j = (this.height - this.backgroundHeight) / 2;
+        context.drawTexture(RenderLayer::getGuiTextured, TEXTURE, i, j, 0.0F, 0.0F, this.backgroundWidth, this.backgroundHeight, 256, 256);
+    }
+}
