--- src/client/java/com/theohenno/narcmod/NarcModClient.java
@@ -1,16 +1,0 @@
-package com.theohenno.narcmod;
-
-import com.theohenno.narcmod.entities.client.DroneEntityClient;
-import com.theohenno.narcmod.screens.ModScreens;
-import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-
-@Environment(EnvType.CLIENT)
-public class NarcModClient implements ClientModInitializer {
-	@Override
-	public void onInitializeClient() {
-		ModScreens.initialize();
-		DroneEntityClient.initialize();
-	}
-}
