--- src/client/java/com/theohenno/narcmod/entities/client/DroneEntityRenderState.java
@@ -1,11 +1,0 @@
-package com.theohenno.narcmod.entities.client;
-
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import net.minecraft.client.render.entity.state.LivingEntityRenderState;
-
-@Environment(EnvType.CLIENT)
-public class DroneEntityRenderState extends LivingEntityRenderState {
-    public DroneEntityRenderState() {
-    }
-}
