--- src/client/resources/narcmod.client.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "com.theohenno.narcmod.mixin.client",
-	"compatibilityLevel": "JAVA_21",
-	"client": [
-		"ExampleClientMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
