+++ src/main/java/com/theohenno/narc_mod/data/DataGenerator.java
@@ -1,0 +1,14 @@
+package com.theohenno.narc_mod.data;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+
+public class DataGenerator implements DataGeneratorEntrypoint {
+
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator generator) {
+        FabricDataGenerator.Pack pack = generator.createPack();
+
+        pack.addProvider(RecipeProvider::new);
+    }
+}
