+++ src/main/java/com/theohenno/narc_mod/data/RecipeProvider.java
@@ -1,0 +1,44 @@
+package com.theohenno.narc_mod.data;
+
+import com.theohenno.narc_mod.blocks.ModBlocks;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
+import net.minecraft.data.recipe.RecipeExporter;
+import net.minecraft.data.recipe.RecipeGenerator;
+import net.minecraft.item.Items;
+import net.minecraft.recipe.book.RecipeCategory;
+import net.minecraft.registry.RegistryWrapper;
+
+import java.util.concurrent.CompletableFuture;
+
+public class RecipeProvider extends FabricRecipeProvider {
+    public RecipeProvider(FabricDataOutput output, CompletableFuture<RegistryWrapper.WrapperLookup> completableFuture) {
+        super(output, completableFuture);
+    }
+
+    @Override
+    public String getName() {
+        return "";
+    }
+
+    @Override
+    protected RecipeGenerator getRecipeGenerator(RegistryWrapper.WrapperLookup registryLookup, RecipeExporter exporter) {
+        return new RecipeGenerator(registryLookup, exporter) {
+
+            @Override
+            public void generate() {
+                createShaped(RecipeCategory.BUILDING_BLOCKS, ModBlocks.COMPUTER, 1)
+                        .pattern("DDD")
+                        .pattern("DRD")
+                        .pattern("DDD")
+
+                        .input('D', Items.DIAMOND)
+                        .input('R', Items.REDSTONE_BLOCK)
+                        .criterion(hasItem(Items.DIAMOND), conditionsFromItem(Items.DIAMOND))
+                        .criterion(hasItem(Items.REDSTONE_BLOCK), conditionsFromItem(Items.REDSTONE_BLOCK))
+                        .offerTo(exporter);
+
+            }
+        };
+    }
+}
