--- src/main/java/com/theohenno/narcmod/DataGenerator.java
@@ -1,15 +1,0 @@
-package com.theohenno.narcmod;
-
-import com.theohenno.narcmod.data.RecipeProvider;
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-
-public class DataGenerator implements DataGeneratorEntrypoint {
-
-    @Override
-    public void onInitializeDataGenerator(FabricDataGenerator generator) {
-        FabricDataGenerator.Pack pack = generator.createPack();
-
-        pack.addProvider(RecipeProvider::new);
-    }
-}
