+++ src/main/resources/assets/narc_mod/blockstates/computer.json
