+++ src/main/resources/assets/narc_mod/icon.png
