+++ src/main/resources/assets/narc_mod/items/computer.json
