+++ src/main/resources/assets/narc_mod/items/suspicious_substance.json
