+++ src/main/resources/assets/narc_mod/lang/en_us.json
@@ -1,0 +1,7 @@
+{
+  "block.narc_mod.computer": "Computer Block",
+  "item.narc_mod.computer": "Computer Block",
+  "item.narc_mod.suspicious_substance": "Suspicious Substance",
+  "itemGroup.narc_mod": "NarcMod",
+  "entity.narc_mod.drone": "Drone"
+}
