+++ src/main/resources/assets/narc_mod/models/block/computer.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:block/cube_all",
+  "textures": {
+    "all": "narc_mod:block/computer"
+  }
+}
