+++ src/main/resources/assets/narc_mod/models/item/suspicious_substance.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:item/generated",
+  "textures": {
+    "layer0": "narc_mod:item/suspicious_substance"
+  }
+}
