+++ src/main/resources/assets/narc_mod/textures/block/computer.png
