+++ src/main/resources/assets/narc_mod/textures/entity/drone/drone.png
