+++ src/main/resources/assets/narc_mod/textures/item/suspicious_substance.png
