--- src/main/resources/assets/narcmod/lang/en_us.json
@@ -1,6 +1,0 @@
-{
-  "block.narcmod.computer": "Computer Block",
-  "item.narcmod.computer": "Computer Block",
-  "item.narcmod.suspicious_substance": "Suspicious Substance",
-  "itemGroup.narcmod": "NarcMod"
-}
