--- src/main/resources/assets/narcmod/models/block/computer.json
@@ -1,6 +1,0 @@
-{
-  "parent": "minecraft:block/cube_all",
-  "textures": {
-    "all": "narcmod:block/computer"
-  }
-}
