--- src/main/resources/assets/narcmod/models/item/suspicious_substance.json
@@ -1,6 +1,0 @@
-{
-  "parent": "minecraft:item/generated",
-  "textures": {
-    "layer0": "narcmod:item/suspicious_substance"
-  }
-}
