--- src/main/resources/narcmod.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "com.theohenno.narcmod.mixin",
-	"compatibilityLevel": "JAVA_21",
-	"mixins": [
-		"ExampleMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
